+++ src/connections/BIDSConn.cpp
@@ -48,9 +48,9 @@ int BIDSConn::setPatchSize(const char * filename)
    int xScalePre = pre->getXScale();
    int xScalePost = post->getXScale();
 
-   int xScale = pow(2, xScalePre);
+   int xScale = (int)pow(2, xScalePre);
    //Convert to bids space, +1 to round up
-   nxp = 1 + 2*(ceil((float)inputParams->value(name, "lateralRadius")/(float)xScale) + ceil((float)2 * (float)inputParams->value(jitterSourceName, "jitter")/(float)xScale));
+   nxp = (int)(1 + 2*(ceil((float)inputParams->value(name, "lateralRadius")/(float)xScale) + ceil((float)2 * (float)inputParams->value(jitterSourceName, "jitter")/(float)xScale)));
 
    status = checkPatchSize(nxp, xScalePre, xScalePost, 'x');
    if( status != PV_SUCCESS) return status;
@@ -58,9 +58,9 @@ int BIDSConn::setPatchSize(const char * filename)
    int yScalePre = pre->getYScale();
    int yScalePost = post->getYScale();
 
-   int yScale = pow(2, yScalePre);
+   int yScale = (int)pow(2, yScalePre);
    //Convert to bids space, +1 to round up
-   nyp = 1 + 2 * (ceil((float)inputParams->value(name, "lateralRadius")/(float)yScale) + ceil((float)2 * (float)inputParams->value(jitterSourceName, "jitter")/(float)yScale));
+   nyp = (int)(1 + 2 * (ceil((float)inputParams->value(name, "lateralRadius")/(float)yScale) + ceil((float)2 * (float)inputParams->value(jitterSourceName, "jitter")/(float)yScale)));
 
    status = checkPatchSize(nyp, yScalePre, yScalePost, 'y');
    if( status != PV_SUCCESS) return status;
