+++ src/connections/LCALIFLateralKernelConn.hpp
@@ -30,7 +30,7 @@ class LCALIFLateralKernelConn: public KernelConn {
       virtual int setParams(PVParams * params); // Really should be protected
 
       virtual int checkpointWrite(const char * cpDir);
-      virtual int checkpointRead(const char * cpDir, float* timef);
+      virtual int checkpointRead(const char * cpDir, double * timef);
 
    protected:
       LCALIFLateralKernelConn();
